fix annoying memory corruption on i386 in readdir due to mismatched prototypes for getdents64...
git-svn-id: svn+ssh://market.scs.stanford.edu/home/u1/asbestos/repo/jos64/trunk@6770 927ebb28-5305-0410-b3ca-c4b640df69b7

+++ lib/fd.cc
@@ -1435,7 +1435,7 @@ ioctl(int fdnum, unsigned long int req, ...) __THROW
 
 extern "C" ssize_t __getdents(int fdnum, struct dirent *buf, size_t nbytes)
 	attribute_hidden;
-extern "C" int64_t __getdents64(int fdnum, struct dirent64 *buf, uint64_t nbytes)
+extern "C" ssize_t __getdents64(int fdnum, struct dirent64 *buf, size_t nbytes)
 	attribute_hidden;
 
 ssize_t
@@ -1444,8 +1444,8 @@ __getdents(int fdnum, struct dirent *buf, size_t nbytes)
     return FD_CALL(fdnum, getdents, buf, nbytes);
 }
 
-int64_t
-__getdents64(int fdnum, struct dirent64 *buf, uint64_t nbytes)
+ssize_t
+__getdents64(int fdnum, struct dirent64 *buf, size_t nbytes)
 {
     return FD_CALL(fdnum, getdents64, buf, nbytes);
 }
